update tick cumulatives on open

+++ contracts/libraries/Position.sol
@@ -24,7 +24,7 @@ library Position {
     function get(
         mapping(bytes32 => Info) storage positions,
         address owner,
-        uint256 id
+        uint112 id
     ) internal view returns (Info memory) {
         return positions[keccak256(abi.encodePacked(owner, id))];
     }
@@ -33,7 +33,7 @@ library Position {
     function set(
         mapping(bytes32 => Info) storage positions,
         address owner,
-        uint256 id,
+        uint112 id,
         Info memory position
     ) internal {
         positions[keccak256(abi.encodePacked(owner, id))] = position;
